+++ src/Ui/AddPage/Advanced.py
@@ -5,7 +5,7 @@ from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QWidget
 
 # 项目内模块导入
-from src.Ui.common.InputCard import SwitchConfigCard, ComboBoxConfigCard, LineEditConfigCard
+from src.Ui.common.input_card import SwitchConfigCard, ComboBoxConfigCard, LineEditConfigCard
 from src.Core.Config.ConfigModel import AdvancedConfig
