+++ src/Ui/common/input_card/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # 项目内模块导入
-from src.Ui.common.InputCard.Item import SwitchItem, LineEditItem
-from src.Ui.common.InputCard.GenericCard import (
+from src.Ui.common.input_card.Item import SwitchItem, LineEditItem
+from src.Ui.common.input_card.generic_card import (
     FolderConfigCard,
     SwitchConfigCard,
     ComboBoxConfigCard,
