+++ src/Ui/common/input_card/generic_card.py
